+++ src/api/clocking/hooks/useClockingFindAllQuery/types.ts
@@ -1,0 +1,4 @@
+import { fetchClockingFindAll } from 'api/clocking/api'
+import { QueryConfig } from 'lib'
+
+export type UseClockingFindAllQueryOptions = QueryConfig<typeof fetchClockingFindAll>