+++ src/constants/api.ts
@@ -8,6 +8,7 @@ const USERACCOUNT_BASE_PATH = '/user-accounts'
 
 export const PERSON_PATH_FIND_ALL = `${PERSON_BASE_PATH}/all`
 export const PERSON_PATH_CREATE = `${PERSON_BASE_PATH}/saveOne`
+export const CLOCKING_PATH_FIND_ALL = `${CLOCKING_BASE_PATH}/all`
 export const CLOCKING_PATH_CREATE = `${CLOCKING_BASE_PATH}/saveOne`
 export const SCHEDULE_ROTATION_PATH_FIND_ALL = `${SCHEDULE_BASE_PATH}/{scheduleId}/rotations`
 export const USER_ACCOUNT_LOGIN = '/backendLoginEndpoint'