+++ src/data/helpers/cleanUrl/cleanUrl.ts
@@ -1,0 +1,4 @@
+export const cleanUrl = (url: string) => ({
+  protocol: (url.match(/^(https?:\/\/)/) || [])[0] || '',
+  url: url.replace(/^(https?:\/\/)/, '').replace(/^www\./, '')
+})