+++ src/routes/index.tsx
@@ -1,5 +1,6 @@
 import OnePerson from '../components/Person/FindById/OnePerson';
 import PersonsList from '../components/Person/FindAll/PersonsList';
+import ClockingsList from '../components/Clocking/FindAll/ClockingsList';
 import AddPerson from 'components/Person/Save/AddPerson';
 import AddClocking from 'components/Clocking/Save/AddClocking';
 import Menu from 'components/Menu';
@@ -10,12 +11,13 @@ import About from 'components/About';
 import NotFound from 'components/NotFound';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import UserAccountsList from 'components/UserAccount/FindAll/UserAccountsList';
+import ExampleComponent from 'components/UserAccount/ExampleComponent';
 
   const AppRoutes = () => {
     return (
       <Router>
         <div style={{ display: 'flex', height: '100vh' }}>
-          <div style={{ width: '250px', backgroundColor: '#f0f0f0', border: '2px solid black' }}> 
+          <div style={{ width: '280px', backgroundColor: '#f0f0f0', border: '2px solid black' }}> 
             <Menu /> 
           </div>
           <div style={{ flex: 1, padding: '20px', overflowY: 'auto' }}> 
@@ -27,8 +29,10 @@ import UserAccountsList from 'components/UserAccount/FindAll/UserAccountsList';
               <Route path='addclocking' element={<AddClocking />} />                
               <Route path='onePerson' element={<OnePerson />} />
               <Route path='personsList' element={<PersonsList />} />
+              <Route path='clockingsList' element={<ClockingsList />} />
               <Route path='userAccountsList' element={<UserAccountsList />} />
               <Route path='logoutAccount' element={<Logout />} />
+              <Route path='exampleComponent' element={<ExampleComponent />} />
               </Route>
               <Route path='*' element={<NotFound />} />
             </Routes>